@@ -2,6 +2,6 @@ namespace PhotoCli.Services.Contracts;
 
 public interface ICsvService
 {
-	Task Report(IEnumerable<Photo> photos, string outputPath, bool isDryRun = false);
-	Task WriteExifDataToCsvOutput(Dictionary<string, ExifData?> photoExifDataByPath, string outputFile);
+	Task CreateCopyReport(IEnumerable<Photo> photos, string outputPath, bool isDryRun = false);
+	Task CreateInfoReport(IEnumerable<Photo> photos, string outputFile);
 }
