@@ -2,6 +2,6 @@ namespace PhotoCli.Services.Contracts;
 
 public interface IDirectoryGrouperService
 {
-	Dictionary<string, List<Photo>> GroupFiles(Dictionary<string, ExifData?> photoExifDataByFilePath, string sourceRootPath, FolderProcessType folderProcessType,
+	Dictionary<string, IReadOnlyCollection<Photo>> GroupFiles(IReadOnlyCollection<Photo> photos, string sourceRootPath, FolderProcessType folderProcessType,
 		GroupByFolderType? groupByFolderType, bool invalidFileFormatGroupedInSubFolder, bool noPhotoDateTimeTakenGroupedInSubFolder, bool noReverseGeocodeGroupedInSubFolder);
 }
