@@ -2,5 +2,5 @@ namespace PhotoCli.Services.Contracts;
 
 public interface IExifDataAppenderService
 {
-	Dictionary<string, ExifData?> ExifDataByPath(IEnumerable<string> photoPaths, out bool allPhotosAreValid, out bool allPhotosHasPhotoTaken, out bool allPhotosHasCoordinate);
+	IReadOnlyCollection<Photo> ExtractExifData(IReadOnlyCollection<Photo> photos, out bool allPhotosAreValid, out bool allPhotosHasPhotoTaken, out bool allPhotosHasCoordinate);
 }
