@@ -2,5 +2,5 @@ namespace PhotoCli.Services.Contracts;
 
 public interface IExifDataGrouperService
 {
-	Dictionary<string, List<Photo>> Group(IEnumerable<Photo> photoInfos, NamingStyle namingStyle);
+	Dictionary<string, List<Photo>> Group(IEnumerable<Photo> photos, NamingStyle namingStyle);
 }
