@@ -2,9 +2,9 @@ namespace PhotoCli.Services.Contracts;
 
 public interface IFileService
 {
-	void Copy(IReadOnlyCollection<Photo> photos, string outputFolder, bool isDryRun);
+	IReadOnlyCollection<Photo> Copy(IReadOnlyCollection<Photo> photos, string outputFolder, bool isDryRun);
 	IReadOnlyCollection<Photo> CopyIfNotExists(IReadOnlyCollection<Photo> photos, string outputFolder, bool isDryRun);
-	Task<bool> VerifyFileIntegrity(IEnumerable<Photo> photos, string outputFolder);
+	Task<bool> VerifyFileIntegrity(IEnumerable<Photo> photos);
 	Task SaveGnuHashFileTree(IEnumerable<Photo> photos, string outputFolder);
-	Task CalculateFileHash(IEnumerable<Photo> photos);
+	Task<IReadOnlyCollection<Photo>> CalculateFileHash(IReadOnlyCollection<Photo> photos);
 }
