@@ -2,6 +2,6 @@ namespace PhotoCli.Services.Contracts;
 
 public interface IFolderRenamerService
 {
-	void RenameByFolderAppendType(IReadOnlyCollection<Photo> orderedPhotoInfos, FolderAppendType folderAppendType, FolderAppendLocationType folderAppendLocationType,
+	IReadOnlyCollection<Photo> RenameByFolderAppendType(IReadOnlyCollection<Photo> orderedPhotos, FolderAppendType folderAppendType, FolderAppendLocationType folderAppendLocationType,
 		string targetRelativeDirectoryPath);
 }
