@@ -2,6 +2,6 @@
 
 public interface IReverseGeocodeFetcherService
 {
-	Task<Dictionary<string, ExifData?>> Fetch(Dictionary<string, ExifData?> exifDataByFilePath);
+	Task<IReadOnlyCollection<Photo>> Fetch(IReadOnlyCollection<Photo> photos);
 	void RateLimitWarning();
 }
