@@ -14,24 +14,24 @@ public class ExifDataAppenderService : IExifDataAppenderService
 		_consoleWriter = consoleWriter;
 	}
 
-	public Dictionary<string, ExifData?> ExifDataByPath(IEnumerable<string> photoPaths, out bool allPhotosAreValid, out bool allPhotosHasPhotoTaken, out bool allPhotosHasCoordinate)
+	public IReadOnlyCollection<Photo> ExtractExifData(IReadOnlyCollection<Photo> photos, out bool allPhotosAreValid, out bool allPhotosHasPhotoTaken, out bool allPhotosHasCoordinate)
 	{
 		_consoleWriter.ProgressStart(ProgressName, _statistics.PhotosFound);
 		var photosAreValid = true;
 		var photosHasPhotoTaken = true;
 		var photosHasCoordinate = true;
 
-		var filePathInfo = new Dictionary<string, ExifData?>();
-		foreach (var photoPath in photoPaths)
+		foreach (var photo in photos)
 		{
-			var photoExifData = _exifParserService.Parse(photoPath, true, true);
-			if (photoExifData == null)
+			var exifData = _exifParserService.Parse(photo.PhotoFile.SourcePath, true, true);
+			if (exifData == null)
 				photosAreValid = false;
-			if (photosHasPhotoTaken && photoExifData?.TakenDate == null)
+			if (photosHasPhotoTaken && exifData?.TakenDate == null)
 				photosHasPhotoTaken = false;
-			if (photosHasCoordinate && photoExifData?.Coordinate == null)
+			if (photosHasCoordinate && exifData?.Coordinate == null)
 				photosHasCoordinate = false;
-			filePathInfo.Add(photoPath, photoExifData);
+			if(exifData != null)
+				photo.SetExifData(exifData);
 			_consoleWriter.InProgressItemComplete(ProgressName);
 		}
 
@@ -39,6 +39,6 @@ public class ExifDataAppenderService : IExifDataAppenderService
 		allPhotosAreValid = photosAreValid;
 		allPhotosHasPhotoTaken = photosHasPhotoTaken;
 		allPhotosHasCoordinate = photosHasCoordinate;
-		return filePathInfo;
+		return photos;
 	}
 }
