@@ -15,52 +15,61 @@ public class FileNamerService : IFileNamerService
 		_logger = logger;
 	}
 
-	public void SetFileName(IReadOnlyCollection<Photo> orderedPhotoInfos, NamingStyle namingStyle, NumberNamingTextStyle numberNamingTextStyle)
+	public IReadOnlyCollection<Photo> SetFileName(IReadOnlyCollection<Photo> orderedPhotos, NamingStyle namingStyle, NumberNamingTextStyle numberNamingTextStyle)
 	{
-		orderedPhotoInfos.ThrowIfNotOrderedByPhotoTakenDate();
+		orderedPhotos.ThrowIfNotOrderedByPhotoTakenDate();
 		if (namingStyle is NamingStyle.Numeric)
 		{
-			_logger.LogTrace("Using numeric numerator for {PhotoCount}", orderedPhotoInfos.Count);
-			using var numberIterator = _sequentialNumberEnumeratorService.NumberIterator(orderedPhotoInfos.Count, numberNamingTextStyle).GetEnumerator();
-			foreach (var photoInfo in orderedPhotoInfos)
+			_logger.LogTrace("Using numeric numerator for {PhotoCount}", orderedPhotos.Count);
+			using var numberIterator = _sequentialNumberEnumeratorService.NumberIterator(orderedPhotos.Count, numberNamingTextStyle).GetEnumerator();
+			foreach (var photoInfo in orderedPhotos)
 			{
 				numberIterator.MoveNext();
-				photoInfo.NewName = numberIterator.Current;
+				photoInfo.SetNewName(numberIterator.Current);
 			}
 		}
 		else
 		{
-			_logger.LogTrace("Grouping {PhotoCount} photos with {NamingStyle}", orderedPhotoInfos.Count, namingStyle);
-			var photosGroupedByDateTime = _exifDataGrouperService.Group(orderedPhotoInfos, namingStyle);
-			foreach (var (groupKey, photoInfos) in photosGroupedByDateTime)
+			_logger.LogTrace("Grouping {PhotoCount} photos with {NamingStyle}", orderedPhotos.Count, namingStyle);
+			var photosGroupedByDateTime = _exifDataGrouperService.Group(orderedPhotos, namingStyle);
+			foreach (var (groupKey, photos) in photosGroupedByDateTime)
 			{
-				if (photoInfos.Count == 1)
+				if (photos.Count == 1)
 				{
-					photoInfos.Single().NewName = groupKey;
+					photos.Single().SetNewName(groupKey);
 					_logger.LogTrace("Single photo grouped with key {GroupKey}. Directly used group key as name, no number iteration needed", groupKey);
 				}
 				else
 				{
-					_logger.LogTrace("{PhotoCount} photos groped with key {GroupKey}. Numerate files with number naming strategy {NumberStrategy}", photoInfos.Count, groupKey, numberNamingTextStyle);
-					using var numberIterator = _sequentialNumberEnumeratorService.NumberIterator(photoInfos.Count, numberNamingTextStyle).GetEnumerator();
-					foreach (var photoInfo in photoInfos)
+					_logger.LogTrace("{PhotoCount} photos groped with key {GroupKey}. Numerate files with number naming strategy {NumberStrategy}", photos.Count, groupKey, numberNamingTextStyle);
+					using var numberIterator = _sequentialNumberEnumeratorService.NumberIterator(photos.Count, numberNamingTextStyle).GetEnumerator();
+					foreach (var photoInfo in photos)
 					{
 						numberIterator.MoveNext();
-						photoInfo.NewName = $"{groupKey}{_options.SameNameNumberSeparator}{numberIterator.Current}";
+						photoInfo.SetNewName($"{groupKey}{_options.SameNameNumberSeparator}{numberIterator.Current}");
 					}
 				}
 			}
 		}
+		return orderedPhotos;
 	}
 
-	public void SetArchiveFileName(IReadOnlyCollection<Photo> orderedPhotoInfos)
+	public IReadOnlyCollection<Photo> SetArchiveFileName(IReadOnlyCollection<Photo> orderedPhotos)
 	{
-		foreach (var photo in orderedPhotoInfos)
+		foreach (var photo in orderedPhotos)
 		{
-			if (photo.PhotoTakenDateTime.HasValue)
-				photo.NewName = photo.PhotoTakenDateTime.Value.ToString(_options.DateTimeFormatWithSeconds) + _options.ArchivePhotoTakenDateHashSeparator + photo.Sha1Hash;
+			if (photo.PhotoFile.Sha1Hash.IsPresent())
+			{
+				if (photo.TakenDateTime.HasValue)
+					photo.SetNewName(photo.TakenDateTime.Value.ToString(_options.DateTimeFormatWithSeconds) + _options.ArchivePhotoTakenDateHashSeparator + photo.PhotoFile.Sha1Hash);
+				else
+					photo.SetNewName(photo.PhotoFile.Sha1Hash);
+			}
 			else
-				photo.NewName = photo.Sha1Hash;
+			{
+				throw new PhotoCliException($"Couldn't set file name for {photo.PhotoFile.SourceFullPath} without Sha1Hash is set");
+			}
 		}
+		return orderedPhotos;
 	}
 }
