@@ -6,7 +6,11 @@ public class PhotoCliException : Exception
 	{
 	}
 
-	public PhotoCliException(string? message) : base(message)
+	public PhotoCliException(string message) : base(message)
+	{
+	}
+
+	public PhotoCliException(string message, Exception innerException) : base(message, innerException)
 	{
 	}
 }
