@@ -4,67 +4,64 @@ public class DirectoryGrouperServiceUnitTests
 {
 	#region Shared
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllValidFilesLocatedInSourceRootShouldGroupedInRootPath = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllValidFilesLocatedInSourceRootShouldGroupedInRootPath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithDay(1) },
+				PhotoInputRootFolder("1", ExifDataFakes.WithDay(1))
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithDay(1), RootTargetRelativePath)
+					]
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceRootPath("f2.jpg"), ExifDataFakes.WithDay(2) }
+				PhotoInputRootFolder("1", ExifDataFakes.WithDay(1)),
+				PhotoInputRootFolder("2", ExifDataFakes.WithDay(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-						SourceRootPhotoInfo("f2.jpg", ExifDataFakes.WithDay(2), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithDay(1), RootTargetRelativePath),
+						GroupedPhotoRootFolder("2", ExifDataFakes.WithDay(2), RootTargetRelativePath)
+					]
 				}
 			}
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> SomeValidFilesSomeInvalidFilesShouldGroupedInRootPath = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> SomeValidFilesSomeInvalidFilesShouldGroupedInRootPath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceRootPath("f4-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f5-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("5-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f4-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfoInvalidFileFormat("f5-root-invalid-file-format.jpg", RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
+						GroupedInvalidPhotoRootFolder("5-root-invalid-file-format", RootTargetRelativePath)
+					]
 				},
 			}
 		},
@@ -76,51 +73,49 @@ public class DirectoryGrouperServiceUnitTests
 
 	#region Without Move Action
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesLocatedInSubFolderShouldGroupedInTheirRelativePath = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesLocatedInSubFolderShouldGroupedInTheirRelativePath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithDay(1) },
+				PhotoInputSubFolder("1", ExifDataFakes.WithDay(1))
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithDay(1), SubPath)
+					]
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceSubPath("f2.jpg"), ExifDataFakes.WithDay(2) }
+				PhotoInputSubFolder("1", ExifDataFakes.WithDay(1)),
+				PhotoInputSubFolder("2", ExifDataFakes.WithDay(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), SubPath),
-						SourceSubPhotoInfo("f2.jpg", ExifDataFakes.WithDay(2), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithDay(1), SubPath),
+						GroupedPhotoSubFolder("2", ExifDataFakes.WithDay(2), SubPath)
+					]
 				}
 			}
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> FilesLocatedInRootAndSubFoldersShouldGroupedInTheirRelativePath = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> FilesLocatedInRootAndSubFoldersShouldGroupedInTheirRelativePath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithDay(2) }
+				PhotoFakes.CreateWithExifData(ExifDataFakes.WithDay(1), SourceRootPath("1-root")),
+				PhotoFakes.CreateWithExifData(ExifDataFakes.WithDay(2), SourceSubPath("2-sub")),
 			},
 			new Dictionary<string, List<Photo>>
 			{
@@ -128,59 +123,57 @@ public class DirectoryGrouperServiceUnitTests
 					RootTargetRelativePath,
 					new List<Photo>
 					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDay(1), RootTargetRelativePath),
 					}
 				},
 				{
 					SubPath,
 					new List<Photo>
 					{
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithDay(2), SubPath),
+						GroupedPhotoSubFolder("2-sub", ExifDataFakes.WithDay(2), SubPath),
 					}
 				}
 			}
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> FilesWithNoExifDataOrInvalidFormatStaysInTheirSourceFolders = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> FilesWithNoExifDataOrInvalidFormatStaysInTheirSourceFolders = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceRootPath("f4-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f5-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
-				{ SourceSubPath("f6-sub.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2) },
-				{ SourceSubPath("f7-sub-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceSubPath("f8-sub-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceSubPath("f9-sub-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceSubPath("f10-sub-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("5-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
+				PhotoInputSubFolder("6-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2)),
+				PhotoInputSubFolder("7-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputSubFolder("8-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputSubFolder("9-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputSubFolder("10-sub-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f4-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfoInvalidFileFormat("f5-root-invalid-file-format.jpg", RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
+						GroupedInvalidPhotoRootFolder("5-root-invalid-file-format", RootTargetRelativePath)
+					]
 				},
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f6-sub.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath),
-						SourceSubPhotoInfo("f7-sub-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), SubPath),
-						SourceSubPhotoInfo("f8-sub-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(), SubPath),
-						SourceSubPhotoInfo("f9-sub-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), SubPath),
-						SourceSubPhotoInfoInvalidFileFormat("f10-sub-invalid-file-format.jpg", SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("6-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath),
+						GroupedPhotoSubFolder("7-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), SubPath),
+						GroupedPhotoSubFolder("8-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(), SubPath),
+						GroupedPhotoSubFolder("9-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), SubPath),
+						GroupedInvalidPhotoSubFolder("10-sub-invalid-file-format", SubPath)
+					]
 				},
 			}
 		},
@@ -193,55 +186,53 @@ public class DirectoryGrouperServiceUnitTests
 	[MemberData(nameof(FilesLocatedInRootAndSubFoldersShouldGroupedInTheirRelativePath))]
 	[MemberData(nameof(FilesWithNoExifDataOrInvalidFormatStaysInTheirSourceFolders))]
 
-	public void SubFoldersPreserveFolderHierarchy_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void SubFoldersPreserveFolderHierarchy_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>>
+			expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.SubFoldersPreserveFolderHierarchy, null, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.SubFoldersPreserveFolderHierarchy, null, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region With Move Action
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHavePhotoTakenDateShouldGroupedInSubFolderIfNoPhotoDateTimeTakenGroupedInSubFolderIsTrue = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHavePhotoTakenDateShouldGroupedInSubFolderIfNoPhotoDateTimeTakenGroupedInSubFolderIsTrue = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceRootPath("f1-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithDay(2) },
-				{ SourceSubPath("f2-sub-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDay(1)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputSubFolder("3-sub", ExifDataFakes.WithDay(2)),
+				PhotoInputSubFolder("4-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDay(1), RootTargetRelativePath)
+					]
 				},
 				{
 					ToolOptionFakes.NoPhotoTakenDateFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName)
+					]
 				},
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithDay(2), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("3-sub", ExifDataFakes.WithDay(2), SubPath)
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName)),
-					}
+					[
+						GroupedPhotoSubFolder("4-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName))
+
+					]
 				},
 			},
 			false,
@@ -250,46 +241,43 @@ public class DirectoryGrouperServiceUnitTests
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHaveGeocodeShouldGroupedInSubFolderIfNoReverseGeocodeGroupedInSubFolderIsTrue = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHaveGeocodeShouldGroupedInSubFolderIfNoReverseGeocodeGroupedInSubFolderIsTrue = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(1) },
-				{ SourceRootPath("f1-root-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(2) },
-				{ SourceSubPath("f2-sub-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithReverseGeocodeSampleId(1)),
+				PhotoInputRootFolder("2-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputSubFolder("3-sub", ExifDataFakes.WithReverseGeocodeSampleId(2)),
+				PhotoInputSubFolder("4-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithReverseGeocodeSampleId(1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithReverseGeocodeSampleId(1), RootTargetRelativePath)
+					]
 				},
 				{
 					ToolOptionFakes.NoAddressFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(), ToolOptionFakes.NoAddressFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(), ToolOptionFakes.NoAddressFolderName)
+					]
 				},
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithReverseGeocodeSampleId(2), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("3-sub", ExifDataFakes.WithReverseGeocodeSampleId(2), SubPath)
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName)),
-					}
+					[
+						GroupedPhotoSubFolder("4-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName))
+
+					]
 				},
 			},
 			false,
@@ -298,94 +286,41 @@ public class DirectoryGrouperServiceUnitTests
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHaveGeocodeAndPhotoTakenDateShouldGroupedInSubFolderIfNoReverseGeocodeGroupedInSubFolderAndNoPhotoDateTimeTakenGroupedInSubFolderIsTrue = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesHasInvalidFormatShouldGroupedInSubFolderIfInvalidFileFormatGroupedInSubFolderIsTrue = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1) },
-				{ SourceRootPath("f1-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2) },
-				{ SourceSubPath("f2-sub-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1)),
+				PhotoInputRootFolder("2-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
+				PhotoInputSubFolder("3-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2)),
+				PhotoInputSubFolder("4-sub-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
-					}
-				},
-				{
-					ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
-							ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName),
-					}
-				},
-				{
-					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath),
-					}
-				},
-				{
-					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName)),
-					}
-				},
-			},
-			false,
-			true,
-			true
-		}
-	};
-
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesHasInvalidFormatShouldGroupedInSubFolderIfInvalidFileFormatGroupedInSubFolderIsTrue = new()
-	{
-		{
-			new Dictionary<string, ExifData?>
-			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1) },
-				{ SourceRootPath("f1-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2) },
-				{ SourceSubPath("f2-sub-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() }
-			},
-			new Dictionary<string, List<Photo>>
-			{
-				{
-					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath)
+					]
 				},
 				{
 					ToolOptionFakes.PhotoFormatInvalidFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfoInvalidFileFormat("f1-root-invalid-file-format.jpg", ToolOptionFakes.PhotoFormatInvalidFolderName),
-					}
+					[
+						GroupedInvalidPhotoRootFolder("2-root-invalid-file-format", ToolOptionFakes.PhotoFormatInvalidFolderName)
+					]
 				},
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("3-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath)
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfoInvalidFileFormat("f2-sub-invalid-file-format.jpg", MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName)),
-					}
+					[
+						GroupedInvalidPhotoSubFolder("4-sub-invalid-file-format", MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName))
+					]
 				},
 			},
 			true,
@@ -394,100 +329,143 @@ public class DirectoryGrouperServiceUnitTests
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesWithNoExifDataOrInvalidFormatGoesIntoItsRelativeSubFolder = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesWithNoExifDataOrInvalidFormatGoesIntoItsRelativeSubFolder = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceRootPath("f4-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f5-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
-				{ SourceSubPath("f6-sub.jpg"), ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2) },
-				{ SourceSubPath("f7-sub-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceSubPath("f8-sub-no-reverse-geocode.jpg"), ExifDataFakes.WithNoReverseGeocode() },
-				{ SourceSubPath("f9-sub-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceSubPath("f10-sub-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("5-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
+
+				PhotoInputSubFolder("6-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2)),
+				PhotoInputSubFolder("7-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputSubFolder("8-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode()),
+				PhotoInputSubFolder("9-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputSubFolder("10-sub-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath)
+					]
 				},
 				{
 					ToolOptionFakes.NoPhotoTakenDateFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName)
+					]
 				},
 				{
 					ToolOptionFakes.NoAddressFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(), ToolOptionFakes.NoAddressFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(), ToolOptionFakes.NoAddressFolderName)
+					]
 				},
 				{
 					ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f4-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
-							ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("4-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
+							ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName)
+
+					]
 				},
 				{
 					ToolOptionFakes.PhotoFormatInvalidFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfoInvalidFileFormat("f5-root-invalid-file-format.jpg", ToolOptionFakes.PhotoFormatInvalidFolderName),
-					}
+					[
+						GroupedInvalidPhotoRootFolder("5-root-invalid-file-format", ToolOptionFakes.PhotoFormatInvalidFolderName)
+					]
 				},
 				{
 					SubPath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f6-sub.jpg", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath),
-					}
+					[
+						GroupedPhotoSubFolder("6-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath)
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f7-sub-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName)),
-					}
+					[
+						GroupedPhotoSubFolder("7-sub-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoPhotoTakenDateFolderName))
+
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f8-sub-no-reverse-geocode.jpg", ExifDataFakes.WithNoReverseGeocode(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName)),
-					}
+					[
+						GroupedPhotoSubFolder("8-sub-no-reverse-geocode", ExifDataFakes.WithNoReverseGeocode(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressFolderName))
+
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f9-sub-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
-							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName)),
-					}
+					[
+						GroupedPhotoSubFolder("9-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName))
+
+					]
 				},
 				{
 					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName),
-					new List<Photo>
-					{
-						SourceSubPhotoInfoInvalidFileFormat("f10-sub-invalid-file-format.jpg", MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName)),
-					}
+				[
+						GroupedInvalidPhotoSubFolder("10-sub-invalid-file-format",
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.PhotoFormatInvalidFolderName))
+					]
 				},
 			},
 			true,
+			true,
+			true
+		}
+	};
+
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FilesNotHaveGeocodeAndPhotoTakenDateShouldGroupedInSubFolderIfNoReverseGeocodeGroupedInSubFolderAndNoPhotoDateTimeTakenGroupedInSubFolderIsTrue = new()
+	{
+		{
+			new []
+			{
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1)),
+				PhotoInputRootFolder("2-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputSubFolder("3-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2)),
+				PhotoInputSubFolder("4-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+			},
+			new Dictionary<string, List<Photo>>
+			{
+				{
+					RootTargetRelativePath,
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDayAndReverseGeocodeSampleId(1, 1), RootTargetRelativePath)
+					]
+				},
+				{
+					ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName,
+					[
+						GroupedPhotoRootFolder("2-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
+							ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName)
+
+					]
+				},
+				{
+					SubPath,
+					[
+						GroupedPhotoSubFolder("3-sub", ExifDataFakes.WithDayAndReverseGeocodeSampleId(2, 2), SubPath)
+					]
+				},
+				{
+					MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName),
+					[
+						GroupedPhotoSubFolder("4-sub-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(),
+							MockFileSystemHelper.RelativePath(SubPath, ToolOptionFakes.NoAddressAndPhotoTakenDateFolderName))
+
+					]
+				},
+			},
+			false,
 			true,
 			true
 		}
@@ -499,10 +477,10 @@ public class DirectoryGrouperServiceUnitTests
 	[MemberData(nameof(FilesNotHaveGeocodeAndPhotoTakenDateShouldGroupedInSubFolderIfNoReverseGeocodeGroupedInSubFolderAndNoPhotoDateTimeTakenGroupedInSubFolderIsTrue))]
 	[MemberData(nameof(FilesHasInvalidFormatShouldGroupedInSubFolderIfInvalidFileFormatGroupedInSubFolderIsTrue))]
 	[MemberData(nameof(FilesWithNoExifDataOrInvalidFormatGoesIntoItsRelativeSubFolder))]
-	public void SubFoldersPreserveFolderHierarchy_SelectedMoveActionForSubFolders_Should_Move_To_Its_Relative_Folder(Dictionary<string, ExifData?> photoExifDataByFilePath,
+	public void SubFoldersPreserveFolderHierarchy_SelectedMoveActionForSubFolders_Should_Move_To_Its_Relative_Folder(IReadOnlyList<Photo> photos,
 		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory, bool invalidFileFormatGroupedInSubFolder, bool noPhotoDateTimeTakenGroupedInSubFolder, bool noReverseGeocodeGroupedInSubFolder)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.SubFoldersPreserveFolderHierarchy, null, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory,
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.SubFoldersPreserveFolderHierarchy, null, photos, expectedGroupedPhotoInfosByRelativeDirectory,
 			invalidFileFormatGroupedInSubFolder, noPhotoDateTimeTakenGroupedInSubFolder, noReverseGeocodeGroupedInSubFolder);
 	}
 
@@ -515,31 +493,31 @@ public class DirectoryGrouperServiceUnitTests
 	[Theory]
 	[MemberData(nameof(AllValidFilesLocatedInSourceRootShouldGroupedInRootPath))]
 	[MemberData(nameof(SomeValidFilesSomeInvalidFilesShouldGroupedInRootPath))]
-	public void Single_GroupFilesByRelativeDirectory_Be_EquivalentTo(Dictionary<string, ExifData?> photoExifDataByFilePath,
-		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void Single_GroupFilesByRelativeDirectory_Be_EquivalentTo(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.Single, null, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.Single, null, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
-	public static TheoryData<Dictionary<string, ExifData?>> SendingFilesLocatedOnSubFoldersOnSingleFolderProcessTypeShouldThrowPhotoOrganizerToolExceptionData = new()
+	public static TheoryData<IReadOnlyList<Photo>> SendingFilesLocatedOnSubFoldersOnSingleFolderProcessTypeShouldThrowPhotoOrganizerToolExceptionData = new()
 	{
-		new Dictionary<string, ExifData?>
+		new List<Photo>
 		{
-			{ SourceSubPath("f-sub.jpg"), ExifDataFakes.WithDay(2) }
+			PhotoFakes.CreateWithExifData(ExifDataFakes.WithDay(1), SourceSubPath("-sub")),
 		},
-		new Dictionary<string, ExifData?>
+
+		new List<Photo>
 		{
-			{ SourceRootPath("f-root.jpg"), ExifDataFakes.WithDay(1) },
-			{ SourceSubPath("f-sub.jpg"), ExifDataFakes.WithDay(2) }
-		}
+			PhotoFakes.CreateWithExifData(ExifDataFakes.WithDay(2), SourceRootPath("-root")),
+			PhotoFakes.CreateWithExifData(ExifDataFakes.WithDay(3), SourceSubPath("-su")),
+		},
 	};
 
 	[Theory]
 	[MemberData(nameof(SendingFilesLocatedOnSubFoldersOnSingleFolderProcessTypeShouldThrowPhotoOrganizerToolExceptionData))]
-	public void Sending_Files_Located_On_Sub_Folders_On_Single_FolderProcessType_Should_Throw_PhotoOrganizerToolException(Dictionary<string, ExifData?> photoExifDataByFilePath)
+	public void Sending_Files_Located_On_Sub_Folders_On_Single_FolderProcessType_Should_Throw_PhotoOrganizerToolException(IReadOnlyList<Photo> photos)
 	{
 		Assert.Throws<PhotoCliException>(() =>
-			GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.Single, null, photoExifDataByFilePath, new Dictionary<string, List<Photo>>())
+			GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.Single, null, photos, new Dictionary<string, List<Photo>>())
 		);
 	}
 
@@ -547,63 +525,60 @@ public class DirectoryGrouperServiceUnitTests
 
 	#region FlattenAllSubFolders
 
-	#region No Group By Folder Type
-
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesLocatedInSubFolderShouldFlattenedIntoRootPath = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesLocatedInSubFolderShouldFlattenedIntoRootPath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithDay(1) },
+				PhotoInputSubFolder("1", ExifDataFakes.WithDay(1)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithDay(1), RootTargetRelativePath)
+					]
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceSubPath("f2.jpg"), ExifDataFakes.WithDay(2) }
+				PhotoInputSubFolder("1", ExifDataFakes.WithDay(1)),
+				PhotoInputSubFolder("2", ExifDataFakes.WithDay(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-						SourceSubPhotoInfo("f2.jpg", ExifDataFakes.WithDay(2), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithDay(1), RootTargetRelativePath),
+						GroupedPhotoSubFolder("2", ExifDataFakes.WithDay(2), RootTargetRelativePath)
+					]
 				}
 			}
 		}
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> FilesLocatedInRootAndSubFoldersShouldFlattenedIntoRootPath = new()
+	#region No Group By Folder Type
+
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> FilesLocatedInRootAndSubFoldersShouldFlattenedIntoRootPath = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-root.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceSubPath("f2-sub.jpg"), ExifDataFakes.WithDay(2) }
+				PhotoInputRootFolder("1-root", ExifDataFakes.WithDay(1)),
+				PhotoInputSubFolder("2-sub", ExifDataFakes.WithDay(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-root.jpg", ExifDataFakes.WithDay(1), RootTargetRelativePath),
-						SourceSubPhotoInfo("f2-sub.jpg", ExifDataFakes.WithDay(2), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("1-root", ExifDataFakes.WithDay(1), RootTargetRelativePath),
+						GroupedPhotoSubFolder("2-sub", ExifDataFakes.WithDay(2), RootTargetRelativePath)
+					]
 				},
 			}
 		}
@@ -614,21 +589,22 @@ public class DirectoryGrouperServiceUnitTests
 	[MemberData(nameof(SomeValidFilesSomeInvalidFilesShouldGroupedInRootPath))]
 	[MemberData(nameof(AllFilesLocatedInSubFolderShouldFlattenedIntoRootPath))]
 	[MemberData(nameof(FilesLocatedInRootAndSubFoldersShouldFlattenedIntoRootPath))]
-	public void FlattenAllSubFolders_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void FlattenAllSubFolders_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>>
+			expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, null, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, null, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region Year
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesThatHasPhotoTakenDateFlattenedAndGroupIntoYearFolders = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesThatHasPhotoTakenDateFlattenedAndGroupIntoYearFolders = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithYear(2001) },
+				PhotoInputRootFolder("1", ExifDataFakes.WithYear(2001)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
@@ -636,34 +612,32 @@ public class DirectoryGrouperServiceUnitTests
 					DateTimeFakes.FormatYear(2001),
 					new List<Photo>
 					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
 					}
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-1.jpg"), ExifDataFakes.WithYear(2001) },
-				{ SourceSubPath("f2.jpg"), ExifDataFakes.WithYear(2002) },
-				{ SourceSubPath("f1-2.jpg"), ExifDataFakes.WithYear(2001) },
+				PhotoInputRootFolder("1-1", ExifDataFakes.WithYear(2001)),
+				PhotoInputSubFolder("2", ExifDataFakes.WithYear(2002)),
+				PhotoInputSubFolder("1-2", ExifDataFakes.WithYear(2001)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.FormatYear(2001),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-1.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
-						SourceSubPhotoInfo("f1-2.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
-					}
+					[
+						GroupedPhotoRootFolder("1-1", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
+						GroupedPhotoSubFolder("1-2", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
+					]
 				},
 				{
 					DateTimeFakes.FormatYear(2002),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2.jpg", ExifDataFakes.WithYear(2002), DateTimeFakes.FormatYear(2002)),
-					}
+					[
+						GroupedPhotoSubFolder("2", ExifDataFakes.WithYear(2002), DateTimeFakes.FormatYear(2002)),
+					]
 				}
 			}
 		}
@@ -671,57 +645,53 @@ public class DirectoryGrouperServiceUnitTests
 
 	[Theory]
 	[MemberData(nameof(AllFilesThatHasPhotoTakenDateFlattenedAndGroupIntoYearFolders))]
-	public void FlattenAllSubFolders_GroupBy_Year_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
-		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void FlattenAllSubFolders_GroupBy_Year_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.Year, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.Year, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region YearMonth
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoYearMonth = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoYearMonth = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithMonth(1) },
+				PhotoInputRootFolder("1", ExifDataFakes.WithMonth(1)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.DirectoryFormatMonth(1),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1)),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1))
+					]
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1-1.jpg"), ExifDataFakes.WithMonth(1) },
-				{ SourceRootPath("f2.jpg"), ExifDataFakes.WithMonth(2) },
-				{ SourceRootPath("f1-2.jpg"), ExifDataFakes.WithMonth(1) },
+				PhotoInputSubFolder("1-1", ExifDataFakes.WithMonth(1)),
+				PhotoInputRootFolder("2", ExifDataFakes.WithMonth(2)),
+				PhotoInputRootFolder("1-2", ExifDataFakes.WithMonth(1)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.DirectoryFormatMonth(1),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1-1.jpg", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1)),
-						SourceRootPhotoInfo("f1-2.jpg", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1)),
-					}
+					[
+						GroupedPhotoSubFolder("1-1", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1)),
+						GroupedPhotoRootFolder("1-2", ExifDataFakes.WithMonth(1), DateTimeFakes.DirectoryFormatMonth(1))
+					]
 				},
 				{
 					DateTimeFakes.DirectoryFormatMonth(2),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2.jpg", ExifDataFakes.WithMonth(2), DateTimeFakes.DirectoryFormatMonth(2)),
-					}
+					[
+						GroupedPhotoRootFolder("2", ExifDataFakes.WithMonth(2), DateTimeFakes.DirectoryFormatMonth(2))
+					]
 				}
 			}
 		}
@@ -729,57 +699,54 @@ public class DirectoryGrouperServiceUnitTests
 
 	[Theory]
 	[MemberData(nameof(AllFilesFlattenedAndGroupIntoYearMonth))]
-	public void FlattenAllSubFolders_GroupBy_YearMonth_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
+	public void FlattenAllSubFolders_GroupBy_YearMonth_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos,
 		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.YearMonth, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.YearMonth, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region YearMonthDay
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoYearMonthDay = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoYearMonthDay = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithDay(1) },
+				PhotoInputSubFolder("1", ExifDataFakes.WithDay(1)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.DirectoryFormatDay(1),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1)),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1))
+					]
 				}
 			}
 		},
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1-1.jpg"), ExifDataFakes.WithDay(1) },
-				{ SourceSubPath("f2.jpg"), ExifDataFakes.WithDay(2) },
-				{ SourceRootPath("f1-2.jpg"), ExifDataFakes.WithDay(1) },
+				PhotoInputRootFolder("1-1", ExifDataFakes.WithDay(1)),
+				PhotoInputSubFolder("2", ExifDataFakes.WithDay(2)),
+				PhotoInputRootFolder("1-2", ExifDataFakes.WithDay(1)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.DirectoryFormatDay(1),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1-1.jpg", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1)),
-						SourceRootPhotoInfo("f1-2.jpg", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1)),
-					}
+					[
+						GroupedPhotoRootFolder("1-1", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1)),
+						GroupedPhotoRootFolder("1-2", ExifDataFakes.WithDay(1), DateTimeFakes.DirectoryFormatDay(1))
+					]
 				},
 				{
 					DateTimeFakes.DirectoryFormatDay(2),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2.jpg", ExifDataFakes.WithDay(2), DateTimeFakes.DirectoryFormatDay(2)),
-					}
+					[
+						GroupedPhotoSubFolder("2", ExifDataFakes.WithDay(2), DateTimeFakes.DirectoryFormatDay(2))
+					]
 				}
 			}
 		}
@@ -787,41 +754,40 @@ public class DirectoryGrouperServiceUnitTests
 
 	[Theory]
 	[MemberData(nameof(AllFilesFlattenedAndGroupIntoYearMonthDay))]
-	public void FlattenAllSubFolders_GroupBy_YearMonthDay_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
+	public void FlattenAllSubFolders_GroupBy_YearMonthDay_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos,
 		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.YearMonthDay, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.YearMonthDay, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region AddressFlat
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoAddressFlat = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoAddressFlat = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f2-1.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(2) },
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(1) },
-				{ SourceSubPath("f2-2.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(2) },
+				PhotoInputSubFolder("2-1", ExifDataFakes.WithReverseGeocodeSampleId(2)),
+				PhotoInputSubFolder("1", ExifDataFakes.WithReverseGeocodeSampleId(1)),
+				PhotoInputSubFolder("2-2", ExifDataFakes.WithReverseGeocodeSampleId(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					ReverseGeocodeFakes.FlatFormatSampleId(1),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithReverseGeocodeSampleId(1), ReverseGeocodeFakes.FlatFormatSampleId(1)),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithReverseGeocodeSampleId(1), ReverseGeocodeFakes.FlatFormatSampleId(1))
+					]
 				},
 				{
 					ReverseGeocodeFakes.FlatFormatSampleId(2),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-1.jpg", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.FlatFormatSampleId(2)),
-						SourceSubPhotoInfo("f2-2.jpg", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.FlatFormatSampleId(2)),
-					}
+					[
+
+						GroupedPhotoSubFolder("2-1", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.FlatFormatSampleId(2)),
+						GroupedPhotoSubFolder("2-2", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.FlatFormatSampleId(2))
+					]
 				},
 			}
 		},
@@ -829,41 +795,38 @@ public class DirectoryGrouperServiceUnitTests
 
 	[Theory]
 	[MemberData(nameof(AllFilesFlattenedAndGroupIntoAddressFlat))]
-	public void FlattenAllSubFolders_GroupBy_AddressFlat_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
-		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void FlattenAllSubFolders_GroupBy_AddressFlat_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.AddressFlat, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.AddressFlat, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region AddressHierarchy
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoAddressHierarchy = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>> AllFilesFlattenedAndGroupIntoAddressHierarchy = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceSubPath("f2-1.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(2) },
-				{ SourceSubPath("f1.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(1) },
-				{ SourceSubPath("f2-2.jpg"), ExifDataFakes.WithReverseGeocodeSampleId(2) },
+				PhotoInputSubFolder("2-1", ExifDataFakes.WithReverseGeocodeSampleId(2)),
+				PhotoInputSubFolder("1", ExifDataFakes.WithReverseGeocodeSampleId(1)),
+				PhotoInputSubFolder("2-2", ExifDataFakes.WithReverseGeocodeSampleId(2)),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					ReverseGeocodeFakes.HierarchyFormatSampleId(1),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f1.jpg", ExifDataFakes.WithReverseGeocodeSampleId(1), ReverseGeocodeFakes.HierarchyFormatSampleId(1)),
-					}
+					[
+						GroupedPhotoSubFolder("1", ExifDataFakes.WithReverseGeocodeSampleId(1), ReverseGeocodeFakes.HierarchyFormatSampleId(1))
+					]
 				},
 				{
 					ReverseGeocodeFakes.HierarchyFormatSampleId(2),
-					new List<Photo>
-					{
-						SourceSubPhotoInfo("f2-1.jpg", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.HierarchyFormatSampleId(2)),
-						SourceSubPhotoInfo("f2-2.jpg", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.HierarchyFormatSampleId(2)),
-					}
+					[
+						GroupedPhotoSubFolder("2-1", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.HierarchyFormatSampleId(2)),
+						GroupedPhotoSubFolder("2-2", ExifDataFakes.WithReverseGeocodeSampleId(2), ReverseGeocodeFakes.HierarchyFormatSampleId(2))
+					]
 				},
 			}
 		},
@@ -871,43 +834,40 @@ public class DirectoryGrouperServiceUnitTests
 
 	[Theory]
 	[MemberData(nameof(AllFilesFlattenedAndGroupIntoAddressHierarchy))]
-	public void FlattenAllSubFolders_GroupBy_AddressHierarchy_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
-		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
+	public void FlattenAllSubFolders_GroupBy_AddressHierarchy_NoMoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos, Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.AddressHierarchy, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory);
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.AddressHierarchy, photos, expectedGroupedPhotoInfosByRelativeDirectory);
 	}
 
 	#endregion
 
 	#region No Exif Data, Invalid File Format - Move Action For Sub Folders
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoNoMoveActionForSubfoldersRootFolder = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoNoMoveActionForSubfoldersRootFolder = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithYear(2001) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f4-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
+				PhotoInputRootFolder("1", ExifDataFakes.WithYear(2001)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode-no-taken-date",  ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("4-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.FormatYear(2001),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001))
+					]
 				},
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfoInvalidFileFormat("f4-root-invalid-file-format.jpg", RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
+						GroupedInvalidPhotoRootFolder("4-root-invalid-file-format", RootTargetRelativePath)
+					]
 				},
 			},
 			false,
@@ -916,39 +876,36 @@ public class DirectoryGrouperServiceUnitTests
 		},
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForInvalidFilesInSubfolders = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForInvalidFilesInSubfolders = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithYear(2001) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f4-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
+				PhotoInputRootFolder("1", ExifDataFakes.WithYear(2001)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("4-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.FormatYear(2001),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001))
+					]
 				},
 				{
 					ToolOptionFakes.PhotoFormatInvalidFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfoInvalidFileFormat("f4-root-invalid-file-format.jpg", ToolOptionFakes.PhotoFormatInvalidFolderName),
-					}
+					[
+						GroupedInvalidPhotoRootFolder("4-root-invalid-file-format", ToolOptionFakes.PhotoFormatInvalidFolderName)
+					]
 				},
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), RootTargetRelativePath)
+					]
 				},
 			},
 			true,
@@ -957,33 +914,31 @@ public class DirectoryGrouperServiceUnitTests
 		},
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForNoPhotoDateTimeTakenInSubFolders = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForNoPhotoDateTimeTakenInSubFolders = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithYear(2001) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f4-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
+				PhotoInputRootFolder("1", ExifDataFakes.WithYear(2001)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("4-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.FormatYear(2001),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001)),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithYear(2001), DateTimeFakes.FormatYear(2001))
+					]
 				},
 				{
 					ToolOptionFakes.NoPhotoTakenDateFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
-						SourceRootPhotoInfoInvalidFileFormat("f4-root-invalid-file-format.jpg", ToolOptionFakes.NoPhotoTakenDateFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), ToolOptionFakes.NoPhotoTakenDateFolderName),
+						GroupedInvalidPhotoRootFolder("4-root-invalid-file-format", ToolOptionFakes.NoPhotoTakenDateFolderName)
+					]
 				},
 			},
 			false,
@@ -992,39 +947,36 @@ public class DirectoryGrouperServiceUnitTests
 		},
 	};
 
-	public static TheoryData<Dictionary<string, ExifData?>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForNoReverseGeocodeGroupedInSubfolders = new()
+	public static TheoryData<IReadOnlyList<Photo>, Dictionary<string, List<Photo>>, bool, bool, bool> FlattenedAndGroupIntoMoveActionForNoReverseGeocodeGroupedInSubfolders = new()
 	{
 		{
-			new Dictionary<string, ExifData?>
+			new []
 			{
-				{ SourceRootPath("f1.jpg"), ExifDataFakes.WithYearAndReverseGeocode(2001) },
-				{ SourceRootPath("f2-root-no-taken-date.jpg"), ExifDataFakes.WithNoPhotoTakenDate() },
-				{ SourceRootPath("f3-root-no-reverse-geocode-no-taken-date.jpg"), ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate() },
-				{ SourceRootPath("f4-root-invalid-file-format.jpg"), ExifDataFakes.WithInvalidFileFormat() },
+				PhotoInputRootFolder("1", ExifDataFakes.WithYearAndReverseGeocode(2001)),
+				PhotoInputRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate()),
+				PhotoInputRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate()),
+				PhotoInputRootFolder("4-root-invalid-file-format", ExifDataFakes.WithInvalidFileFormat()),
 			},
 			new Dictionary<string, List<Photo>>
 			{
 				{
 					DateTimeFakes.FormatYear(2001),
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f1.jpg", ExifDataFakes.WithYearAndReverseGeocode(2001), DateTimeFakes.FormatYear(2001)),
-					}
+					[
+						GroupedPhotoRootFolder("1", ExifDataFakes.WithYearAndReverseGeocode(2001), DateTimeFakes.FormatYear(2001))
+					]
 				},
 				{
 					ToolOptionFakes.NoAddressFolderName,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f3-root-no-reverse-geocode-no-taken-date.jpg", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), ToolOptionFakes.NoAddressFolderName),
-						SourceRootPhotoInfoInvalidFileFormat("f4-root-invalid-file-format.jpg", ToolOptionFakes.NoAddressFolderName),
-					}
+					[
+						GroupedPhotoRootFolder("3-root-no-reverse-geocode-no-taken-date", ExifDataFakes.WithNoReverseGeocodeAndNoTakenDate(), ToolOptionFakes.NoAddressFolderName),
+						GroupedInvalidPhotoRootFolder("4-root-invalid-file-format", ToolOptionFakes.NoAddressFolderName)
+					]
 				},
 				{
 					RootTargetRelativePath,
-					new List<Photo>
-					{
-						SourceRootPhotoInfo("f2-root-no-taken-date.jpg", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath),
-					}
+					[
+						GroupedPhotoRootFolder("2-root-no-taken-date", ExifDataFakes.WithNoPhotoTakenDate(), RootTargetRelativePath)
+					]
 				},
 			},
 			false,
@@ -1038,10 +990,10 @@ public class DirectoryGrouperServiceUnitTests
 	[MemberData(nameof(FlattenedAndGroupIntoMoveActionForInvalidFilesInSubfolders))]
 	[MemberData(nameof(FlattenedAndGroupIntoMoveActionForNoPhotoDateTimeTakenInSubFolders))]
 	[MemberData(nameof(FlattenedAndGroupIntoMoveActionForNoReverseGeocodeGroupedInSubfolders))]
-	public void FlattenAllSubFolders_GroupBy_MoveActionForSubFolders_Should_Be_EquivalentTo_Expected(Dictionary<string, ExifData?> photoExifDataByFilePath,
+	public void FlattenAllSubFolders_GroupBy_MoveActionForSubFolders_Should_Be_EquivalentTo_Expected(IReadOnlyList<Photo> photos,
 		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory, bool invalidFileFormatGroupedInSubFolder, bool noPhotoDateTimeTakenGroupedInSubFolder, bool noReverseGeocodeGroupedInSubFolder)
 	{
-		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.Year, photoExifDataByFilePath, expectedGroupedPhotoInfosByRelativeDirectory,
+		GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType.FlattenAllSubFolders, GroupByFolderType.Year, photos, expectedGroupedPhotoInfosByRelativeDirectory,
 			invalidFileFormatGroupedInSubFolder, noPhotoDateTimeTakenGroupedInSubFolder, noReverseGeocodeGroupedInSubFolder);
 	}
 
@@ -1051,13 +1003,13 @@ public class DirectoryGrouperServiceUnitTests
 
 	#region Helpers
 
-	private static void GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType folderProcessType, GroupByFolderType? groupByFolderType, Dictionary<string, ExifData?> photoExifDataByFilePath,
+	private static void GroupFilesByRelativeDirectoryBeEquivalentTo(FolderProcessType folderProcessType, GroupByFolderType? groupByFolderType, IReadOnlyList<Photo> photos,
 		Dictionary<string, List<Photo>> expectedGroupedPhotoInfosByRelativeDirectory, bool invalidFileFormatGroupedInSubFolder = false, bool noPhotoDateTimeTakenGroupedInSubFolder = false, bool noReverseGeocodeGroupedInSubFolder = false)
 	{
 		var fileSystem = new MockFileSystem();
 		var sut = new DirectoryGrouperService(fileSystem, ToolOptionFakes.Create(), NullLogger<DirectoryGrouperService>.Instance, ConsoleWriterFakes.Valid());
 
-		var actualGroupedPhotoPathsByRelativeDirectory = sut.GroupFiles(photoExifDataByFilePath, SourcePath, folderProcessType, groupByFolderType,
+		var actualGroupedPhotoPathsByRelativeDirectory = sut.GroupFiles(photos, SourcePath, folderProcessType, groupByFolderType,
 			invalidFileFormatGroupedInSubFolder, noPhotoDateTimeTakenGroupedInSubFolder, noReverseGeocodeGroupedInSubFolder);
 
 		actualGroupedPhotoPathsByRelativeDirectory.Should().BeEquivalentTo(expectedGroupedPhotoInfosByRelativeDirectory);
@@ -1067,39 +1019,58 @@ public class DirectoryGrouperServiceUnitTests
 
 	#region Fakes
 
-	private static Photo SourceRootPhotoInfo(string fileName, ExifData? exifData, string targetRelativeDirectoryPath)
+	#region Root Folder
+
+	private static Photo PhotoInputRootFolder(string fileName, ExifData? exifData)
 	{
-		return PhotoFakes.WithValidFilePathAndExifData(SourcePath, fileName, exifData, targetRelativeDirectoryPath);
+		return PhotoFakes.WithSourcePathAndExifData(SourceRootPath(fileName), exifData);
 	}
 
-	private static Photo SourceSubPhotoInfo(string fileName, ExifData? exifData, string targetRelativeDirectoryPath)
+	private static Photo GroupedPhotoRootFolder(string fileName, ExifData? exifData, string targetRelativeDirectoryPath)
 	{
-		return PhotoFakes.WithValidFilePathAndExifData($"{SourcePath}/{SubPath}", fileName, exifData, targetRelativeDirectoryPath);
+		return PhotoFakes.WithValidFilePathAndExifData($"/{SourcePath}", $"{fileName}.{PhotoExtension}", exifData, targetRelativeDirectoryPath);
 	}
 
-	private static Photo SourceRootPhotoInfoInvalidFileFormat(string fileName, string targetRelativeDirectoryPath)
+	private static Photo GroupedInvalidPhotoRootFolder(string fileName, string targetRelativeDirectoryPath)
 	{
-		return PhotoFakes.WithValidFilePathInvalidFileFormat(SourcePath, fileName, targetRelativeDirectoryPath);
-	}
-
-	private static Photo SourceSubPhotoInfoInvalidFileFormat(string fileName, string targetRelativeDirectoryPath)
-	{
-		return PhotoFakes.WithValidFilePathInvalidFileFormat($"{SourcePath}/{SubPath}", fileName, targetRelativeDirectoryPath);
+		return PhotoFakes.WithValidFilePathInvalidFileFormat($"/{SourcePath}", $"{fileName}.{PhotoExtension}", targetRelativeDirectoryPath);
 	}
 
 	private static string SourceRootPath(string fileName)
 	{
-		return MockFileSystemHelper.Path($"/{SourcePath}/{fileName}");
+		return MockFileSystemHelper.Path($"/{SourcePath}/{fileName}.{PhotoExtension}");
+	}
+
+	#endregion
+
+	#region Sub Folder
+
+	private static Photo PhotoInputSubFolder(string fileName, ExifData? exifData)
+	{
+		return PhotoFakes.WithSourcePathAndExifData(SourceSubPath(fileName), exifData);
+	}
+
+	private static Photo GroupedPhotoSubFolder(string fileName, ExifData? exifData, string targetRelativeDirectoryPath)
+	{
+		return PhotoFakes.WithValidFilePathAndExifData($"/{SourcePath}/{SubPath}", $"{fileName}.{PhotoExtension}", exifData, targetRelativeDirectoryPath);
+	}
+
+	private static Photo GroupedInvalidPhotoSubFolder(string fileName, string targetRelativeDirectoryPath)
+	{
+		return PhotoFakes.WithValidFilePathInvalidFileFormat($"/{SourcePath}/{SubPath}", $"{fileName}.{PhotoExtension}", targetRelativeDirectoryPath);
 	}
 
 	private static string SourceSubPath(string fileName)
 	{
-		return MockFileSystemHelper.Path($"/{SourcePath}/{SubPath}/{fileName}");
+		return MockFileSystemHelper.Path($"/{SourcePath}/{SubPath}/{fileName}.{PhotoExtension}");
 	}
+
+	#endregion
 
 	private const string SourcePath = "source-path";
 	private const string SubPath = "sub-path";
 	private const string RootTargetRelativePath = "";
+	private const string PhotoExtension = "jpg";
 
 	#endregion
 }
